RIN-623 - Adds UI framework
- Adds React + Next.js for server side rendering ease.
- Combines server.js and app.js to server.js per Next.js guidelines.
- Adds test index page.
- Adds node_modules, .next to gitignore.
