--- index.html
@@ -1,1 +1,0 @@
-sample websocket
