+++ pages/index.js
@@ -1,0 +1,3 @@
+export default () => (
+  <div>Welcome to next.js! oh snap</div>
+)
